ai-service/requirements.txt
Flask==3.0.3
Flask-Limiter==4.1.1
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.1
groq==0.9.0
